requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
aiosmtplib==2.0.2
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncmy==0.2.9
attrs==25.1.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.8.1
boto3==1.34.98
botocore==1.34.98
cachetools==5.5.0
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.11.1
databases==0.9.0
Deprecated==1.2.14
dkimpy==1.1.8
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
et_xmlfile==2.0.0
eventlet==0.36.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-csrf-protect==0.3.4
fastapi-mail==1.4.1
frozenlist==1.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_resources==6.4.0
iniconfig==2.0.0
isodate==0.7.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.3.0
Mako==1.3.2
MarkupSafe==2.1.5
more-itertools==10.5.0
multidict==6.1.0
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.0
pandas==2.2.3
passlib==1.7.4
phonenumbers==8.13.32
pillow==10.4.0
pluggy==1.4.0
premailer==3.10.0
prompt-toolkit==3.0.43
propcache==0.2.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyMySQL==1.1.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
uuid==1.30
uvicorn==0.28.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.18.3
